requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
psycopg2==2.9.7
psycopg2-binary==2.9.7
flask-migrate== 4.0.4
python-ulid==1.1.0
python-dotenv==1.0.0
alembic==1.12.0
pytz==2023.3.post1